+++ tests/conftest.py
@@ -261,6 +261,7 @@ def case_obj(request, parsed_case):
     logger.info("Create a case obj")
     case = parsed_case
     case['_id'] = parsed_case['case_id']
+    case['owner'] = parsed_case['owner']
     case['created_at'] = parsed_case['analysis_date']
     case['dynamic_gene_list'] = {}
     case['genome_version'] = None